rename options
