+++ localcmd/doc.go
@@ -1,3 +1,3 @@
 // Package localcommand provides an implementation of webtty.Slave
 // that launches a local command with a PTY.
-package localcommand
+package localcmd
